requirements.txt
numpy
pandas
faiss-cpu
sentence-transformers
transformers
scikit-learn
tqdm
python-dotenv
